Reworked the plugin to use Hudson's remote access API instead of an RSS feed to get the build-info. This fixes a number of issues:  * the timeline now always shows all the builds in the selected range;    previously only the recent builds were being shown.  * displaying of modules works again (though enabling this may slow down    the operation significantly).  * builds are displayed in the timeline at the point they completed    instead of at the point they started
This also removes the dependency on the external 'feedparser' library.

One issue with this change is that it may be noticeably slower if you
have many builds in your history.

Upgrade note: the 'feed_url' option has been renamed 'job_url' and now
points to the top-level hudson url or to a job url; and the
'display_subprojects' option has been renamed to 'display_modules'.