Handle WBS and dependencies for tickets whose parents don't match query.
If a ticket depended on a ticket which wasn't on the chart, no
dependencies were drawn.

I first encountered this with "owner=somebody" because a ticket that
matched that was the child of a ticket owned by someone else but it
could happen in general.

Increment version number for this fix.