AnnouncerPlugin: Switch permission check from !PermissionSystem to !PermissionCache, refs #5825 and #8458.
This has been requested for playing nicer with plugins that implement
IPermissionPolicy but do not deal with `check_permission` methode,
if user permission cache is missing as additional argument there.
Currently this applies i.e. to PrivateTicketsPlugin, so the same
workaround has been implemented for TracHoursPlugin by Ryan Ollos before
(see #5825 for details).