11.0.0.6: Skip permission check on resource containing the `TOC`
Permission must have already been granted on this resource
if the macro is being executed. An effect of this change is that
`TOC` can be used in resources other than wiki pages, such as
blog posts, as long as the `TOC` is only created for the post
containing the `TOC` macro.

Fixes #12914.