WikiExtrasPlugin: make the plugin compatible with Trac's jinja2 branch.
This is mainly about replacing the direct Genshi imports with indirect
ones.  By going through `trac.util.html`, we can work with a broad
range of older Trac versions (at least 0.12.x to 1.2.x) and yet be
compatible with future versions which won't require Genshi anymore.

The one compatibility fix which was not trivial was a
re-implementation of the `tracwikiextras.util.sanitize_attrib`
function in terms of the (new) `TracHTMLSanitizer.sanitize_attrs`
method.